build: migrate to tsdown and fix bundling issue
